Spandan2024_testing/Spandan-Backend/spandan_backend/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.14.0
django-filter==22.1
django-jsonfield==1.4.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
idna==3.4
importlib-metadata==6.0.0
jsonfield==3.1.0
Markdown==3.4.1
PyJWT==2.6.0
pymongo==4.3.3
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
zipp==3.15.0
django-otp==1.3.0
mysqlclient==2.2.4